+++ app/src/main/java/com/github/damontecres/wholphin/services/SeerrApi.kt
@@ -25,7 +25,8 @@ class SeerrApi(
     fun update(
         baseUrl: String,
         apiKey: String?,
+        sessionCookie: String? = null,
     ) {
-        api = SeerrApiClient(createSeerrApiUrl(baseUrl), apiKey, okHttpClient)
+        api = SeerrApiClient(createSeerrApiUrl(baseUrl), apiKey, okHttpClient, sessionCookie)
     }
 }